# v2_api/api_docs.yaml
swagger: '2.0'
info:
  version: '2.0'
  title: Book-A-Meal
  description: >-
    Book-A-Meal is an application that allows customers to make food orders and helps the food vendor know what the customers want to eat
  license:
    name: MIT
    url: 'https://github.com/SrMoffat/Book-A-Meal/blob/master/LICENSE'
basePath: /api/v1
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
paths:
...
  /meals:
    x-summary: Meals
    post:
      summary: Admin(Caterer) can add a new meal
      description: Logged in admin (Caterer) can add a new meal
      parameters:
        - in: body
          name: Meals
          description: Meal details
          schema:
            $ref: '#/definitions/AddMeal'
        - name: access-token
          in: header
          description: Auth token in header
          required: true
          type: string
      responses:
        '201':
          description: 'Meal has been successfully created!'
        '400':
          description: 'Meal already exists'
      tags:
        - Meals
    get:
      summary: Admin(Caterer) can get all meals
      description: Logged in admin (Caterer) can get all current meals
      parameters:
        - name: access-token
          in: header
          description: Auth token in header
          required: true
          type: string
      responses:
        '200':
          description: 'All existing meals'
        '404':
          description: 'There are no existing meals yet'
      tags:
        - Meals
...
definitions:
...
  AddMeal:
    type: object
    properties:
      name:
        type: string
      category:
        type: string
      price:
        type: number
      image_url:
        type: string
      description:
        type: string
    example:
      name: Chicken Mamboleo
      price: 200
      category: Lunch
      description: Chicken drumsticks
...